+++ docs/release-notes.md
@@ -16,6 +16,7 @@ These changes have not been released yet.
   * Fixed 'received message' logs shown in non-developer mode.
   * Fixed some assets not updated when you switch language to English.
   * Fixed lag in some cases due to incorrect asset caching when playing in non-English.
+  * Fixed lag when a mod invalidates many NPC portraits/sprites at once.
 
 * For modders:
   * Added support for content pack translations.